[Pdm2-6] Atomize CharacterContent into several components
